Update to build Docker Image with ubuntu user
This commit add the logic to build the docker image with the `ubuntu`
user with UID and GID as 1000:1000 respectively, instead of creating the
user with the default user and GID:UID present at the time of docker
build.

This will ensure consistency in docker image builds, as the runners
used in the github have the same `ubuntu` user with 1000:1000 as UID and GID.
This will help to save time during the workflows as the workflows will now not be
requiring to build the image everytime instead build it once and
store it in the registry so that the other jobs can just pull the docker
image and use it, saving time and computation power of the runners.

Signed-off-by: Hasibur Rahman Mohammed <hasiburr@qti.qualcomm.com>